Resubscribe candles when YES/NO is switched
