+++ src/pages/trader-page/components/candles-webSocket-listener/useCandleMarketsSubscribe.ts
@@ -1,8 +1,10 @@
 import { useAtom, useAtomValue, useSetAtom } from 'jotai';
-import { useEffect, useRef } from 'react';
+import { useEffect, useRef, useState } from 'react';
 
-import { selectedPerpetualAtom } from 'store/pools.store';
+import { orderBlockAtom } from 'store/order-block.store';
+import { selectedPerpetualAtom, selectedPerpetualDataAtom } from 'store/pools.store';
 import { candlesDataReadyAtom, newCandleAtom, selectedPeriodAtom } from 'store/tv-chart.store';
+import { OrderBlockE } from 'types/enums';
 
 import { createPairWithPeriod } from './helpers/createPairWithPeriod';
 import { subscribingCheckAtom, unsubscribeLostCandleAtom } from './subscribingCheckAtom';
@@ -13,15 +15,43 @@ interface UseCandleMarketsSubscribePropsI {
 }
 
 export const useCandleMarketsSubscribe = ({ isConnected, send }: UseCandleMarketsSubscribePropsI) => {
+  const orderBlock = useAtomValue(orderBlockAtom);
   const selectedPeriod = useAtomValue(selectedPeriodAtom);
   const selectedPerpetual = useAtomValue(selectedPerpetualAtom);
+  const selectedPerpetualData = useAtomValue(selectedPerpetualDataAtom);
   const setNewCandle = useSetAtom(newCandleAtom);
   const setCandlesDataReady = useSetAtom(candlesDataReadyAtom);
   const subscribingCheck = useSetAtom(subscribingCheckAtom);
   const [unsubscribeLostCandle, setUnsubscribeLostCandle] = useAtom(unsubscribeLostCandleAtom);
 
+  const [triggerResubscribe, setTriggerResubscribe] = useState(false);
+
   const wsConnectedStateRef = useRef(false);
   const topicRef = useRef('');
+  const latestOrderBlockRef = useRef<OrderBlockE | null>(null);
+  const resubscribeRef = useRef(false);
+
+  const isPredictionMarket = selectedPerpetualData?.isPredictionMarket ?? false;
+
+  // This use effect is used to resubscribe
+  useEffect(() => {
+    if (isPredictionMarket) {
+      if (latestOrderBlockRef.current === null) {
+        latestOrderBlockRef.current = orderBlock;
+        return;
+      }
+
+      // Check if orderBlock is changed
+      if (latestOrderBlockRef.current === orderBlock) {
+        return;
+      }
+
+      latestOrderBlockRef.current = orderBlock;
+
+      // Just ask to resubscribe
+      setTriggerResubscribe((prev) => !prev);
+    }
+  }, [isPredictionMarket, orderBlock]);
 
   useEffect(() => {
     if (selectedPerpetual && isConnected) {
@@ -32,11 +62,12 @@ export const useCandleMarketsSubscribe = ({ isConnected, send }: UseCandleMarket
       wsConnectedStateRef.current = true;
 
       const topicInfo = createPairWithPeriod(selectedPerpetual, selectedPeriod);
-      if (topicInfo !== topicRef.current) {
+      if (topicInfo !== topicRef.current || triggerResubscribe !== resubscribeRef.current) {
         if (topicRef.current) {
           send(JSON.stringify({ type: 'unsubscribe', topic: topicRef.current }));
         }
 
+        resubscribeRef.current = triggerResubscribe;
         topicRef.current = topicInfo;
         send(
           JSON.stringify({
@@ -61,7 +92,16 @@ export const useCandleMarketsSubscribe = ({ isConnected, send }: UseCandleMarket
       wsConnectedStateRef.current = false;
       topicRef.current = '';
     }
-  }, [selectedPerpetual, selectedPeriod, setNewCandle, setCandlesDataReady, isConnected, send, subscribingCheck]);
+  }, [
+    selectedPerpetual,
+    selectedPeriod,
+    setNewCandle,
+    setCandlesDataReady,
+    isConnected,
+    send,
+    subscribingCheck,
+    triggerResubscribe,
+  ]);
 
   useEffect(() => {
     if (unsubscribeLostCandle !== '') {
